fix: dispatch env init to sub modules init
