+++ deploy/aws/fastapi-daily-chat-bot/app/requirements.txt
@@ -1,97 +1,20 @@
-aenum==3.1.15
-aiofiles==24.1.0
-aiohappyeyeballs==2.3.5
-aiohttp==3.10.3
-aiosignal==1.3.1
-annotated-types==0.7.0
-anyio==4.4.0
-apipeline==0.1.7
-attrs==23.2.0
-cattrs==23.2.3
-certifi==2024.7.4
-cffi==1.17.0
-charset-normalizer==3.3.2
-click==8.1.7
-constructs==10.3.0
-daily-python==0.10.1
-dataclasses-json==0.6.7
-deepgram-sdk==3.4.0
-deprecation==2.1.0
-distro==1.9.0
-einops==0.8.0
-einx==0.3.0
-encodec==0.1.1
 fastapi==0.112.0
-filelock==3.15.4
-frozendict==2.4.4
-frozenlist==1.4.1
-fsspec==2024.6.1
-future==1.0.0
-greenlet==3.0.3
-h11==0.14.0
-httpcore==1.0.5
-httpx==0.27.0
-huggingface-hub==0.24.5
-idna==3.7
-importlib_resources==6.4.0
-Jinja2==3.1.4
-jiter==0.5.0
-jsii==1.102.0
-jsonpatch==1.33
-jsonpointer==3.0.0
-langchain==0.2.13
-langchain-community==0.2.12
-langchain-core==0.2.30
-langchain-openai==0.1.21
-langchain-text-splitters==0.2.2
-langsmith==0.1.99
-llvmlite==0.43.0
-mangum==0.17.0
-MarkupSafe==2.1.5
-marshmallow==3.21.3
-mpmath==1.3.0
-multidict==6.0.5
-mypy-extensions==1.0.0
-networkx==3.3
-numba==0.60.0
-numpy==1.26.4
-openai==1.40.6
-orjson==3.10.7
-packaging==24.1
-pillow==10.4.0
-publication==0.0.3
-PyAudio==0.2.14
-pybase16384==0.3.7
-pycparser==2.22
-pydantic==2.8.2
-pydantic_core==2.20.1
-pydub==0.25.1
-pyloudnorm==0.1.1
-python-dateutil==2.9.0.post0
-python-dotenv==1.0.1
-PyYAML==6.0.2
-regex==2024.7.24
 requests==2.32.3
-safetensors==0.4.4
-scipy==1.14.0
-six==1.16.0
-sniffio==1.3.1
-SQLAlchemy==2.0.32
-starlette==0.37.2
-sympy==1.13.2
-tenacity==8.5.0
-tiktoken==0.7.0
-tokenizers==0.19.1
-torch==2.2.2
-torchaudio==2.2.2
-tqdm==4.66.5
-transformers==4.44.0
-typeguard==2.13.3
-typing-inspect==0.9.0
-typing_extensions==4.12.2
-urllib3==2.2.2
-uvicorn==0.30.5
-vector-quantize-pytorch==1.15.6
-vocos==0.1.0
+apipeline==0.1.7
+python-dotenv==1.0.1
+pydub==0.25.1
+openai==1.40.6
+pillow==10.4.0
 websockets==12.0
-yarl==1.9.4
+aiohttp==3.10.3
+daily-python==0.10.1
+scipy==1.14.0
+pyloudnorm==0.1.1
+numpy==2.0.1
+torch==2.2.2
+langchain==0.2.13
+langchain-openai==0.1.21
+langchain-community==0.2.12
+deepgram-sdk==3.4.0
+mangum==0.17.0
+uvicorn==0.30.6
