+++ src/common/types.py
@@ -50,6 +50,9 @@ CHANNELS = 1
 RATE = 16000
 SAMPLE_WIDTH = 2
 
+# llm sys default prompt
+DEFAULT_SYSTEM_PROMPT = "你是一个中国人,一名中文助理，请用中文简短回答，回答限制在1-5句话内。要友好、乐于助人且简明扼要。保持对话简短而甜蜜。只用纯文本回答，不要包含链接或其他附加内容。不要回复计算机代码以及数学公式。"
+
 
 @dataclass
 class SessionCtx:
@@ -109,6 +112,35 @@ SILENCE_TIMEOUT_S = 10
 MIC_IDX = 1
 # 2^(16-1)
 INT16_MAX_ABS_VALUE = 32768.0
+
+# pyaudio format
+#   >>> print(int(pyaudio.paInt16))
+#   8
+PYAUDIO_PAINT16 = 8
+#   >>> print(int(pyaudio.paInt24))
+#   4
+PYAUDIO_PAINT24 = 4
+#   >>> print(int(pyaudio.paInt32))
+#   2
+PYAUDIO_PAINT32 = 2
+#   >>> print(int(pyaudio.paFloat32))
+#   1
+PYAUDIO_PAFLOAT32 = 1
+#   >>> print(int(pyaudio.paInt8))
+#   16
+PYAUDIO_PAINT8 = 16
+#   >>> print(int(pyaudio.paUInt8))
+#   32
+PYAUDIO_PAUINT8 = 32
+#   >>> print(int(pyaudio.paCustomFormat))
+#   65536
+PYAUDIO_PACUSTOMFORMAT = 65536
+
+# PortAudio Callback Return Codes
+
+PYAUDIO_PACONTINUE = 0  #: There is more audio data to come
+PYAUDIO_PACOMPLETE = 1  #: This was the last block of audio data
+PYAUDIO_PAABORT = 2   #: An error ocurred, stop playback/recording
 
 
 @dataclass
