+++ src/modules/speech/audio_stream/pyaudio_stream.py
@@ -1,9 +1,10 @@
 import logging
+
 import pyaudio
 
 from src.common.interface import IAudioStream
 from src.common.factory import EngineClass
-from src.types.speech.audio_stream.pyaudio import AudioStreamInfo, PyAudioStreamArgs
+from src.types.speech.audio_stream import AudioStreamInfo, PyAudioStreamArgs
 
 
 class PyAudioStream(EngineClass, IAudioStream):
