+++ test/modules/speech/recorder/test_record.py
@@ -4,14 +4,16 @@ import asyncio
 
 import unittest
 
-from src.cmd.init import Env
+from src.modules.speech.audio_stream import AudioStreamEnvInit
 from src.common.interface import IAudioStream, IRecorder
 from src.common.logger import Logger
 from src.common.factory import EngineFactory, EngineClass
 from src.common.session import Session
 from src.common.utils import audio_utils
 from src.common.types import SessionCtx, MODELS_DIR, RECORDS_DIR
-import src.modules.speech
+
+import src.modules.speech.recorder.rms_record
+import src.modules.speech.recorder.vad_record
 
 r"""
 python -m unittest test.modules.speech.recorder.test_record.TestRMSRecorder.test_record
@@ -72,7 +74,7 @@ class TestRMSRecorder(unittest.TestCase):
         self.kwargs["is_stream_callback"] = bool(os.getenv('IS_STREAM_CALLBACK', None))
         self.recorder: IRecorder | EngineClass = EngineFactory.get_engine_by_tag(
             EngineClass, self.tag, **self.kwargs)
-        self.audio_in_stream: IAudioStream | EngineClass = Env.initAudioInStreamEngine()
+        self.audio_in_stream: IAudioStream | EngineClass = AudioStreamEnvInit.initAudioInStreamEngine()
         self.recorder.set_in_stream(self.audio_in_stream)
         self.recorder.open()
         self.session = Session(**SessionCtx(
